+++ src/parser/expect/DFA/state.ts
@@ -28,23 +28,18 @@ export class State<T> extends BaseState<
 
   /**
    * Try to use lexer to yield an ASTNode with type and/or content needed by a candidate.
+   * Return all the possible results.
    */
   tryLex(
     lexer: ILexer,
     followSets: ReadonlyMap<string, GrammarSet>
-  ): ASTNode<T> | null {
-    // try to use lexer to yield an ASTNode with specific type and/or content
-    for (let i = 0; i < this.candidates.length; ++i) {
-      const node = this.candidates[i].tryLex(lexer, followSets);
-      if (node !== null) {
-        // for now we only consider the first candidate that can lex the input
-        return node;
-        // TODO: what if multiple candidates can lex the input?
-      }
-    }
-
-    // no candidate can lex the input, return null
-    return null;
+  ): { node: ASTNode<T>; lexer: ILexer }[] {
+    const res: { node: ASTNode<T>; lexer: ILexer }[] = [];
+    this.candidates.map((c) => {
+      const l = lexer.clone(); // each candidate should have its own lexer to avoid side effect
+      res.push(...c.tryLex(l, followSets));
+    });
+    return res;
   }
 
   /** Traverse all candidates to try to reduce. */