+++ CHANGELOG.md
@@ -34,6 +34,7 @@
       - Add `DFA.calculateAllStates` to calculate all state transitions ahead of time and cache them.
   - ELR Parser: Expectational LR Parser
     - Parser will actively use Lexer to lex input string to token according to the grammar rules.
+    - Re-Lex: If the lexed token can't be accepted, the parser will try to restore & re-lex input.
 - Remove `Manager` since `Parser.parse` already accept string as input and keep a state.
 - Optimize document / comments.
   - Generate documentation with [typedoc](https://typedoc.org/).